requirements.txt
# requirements.txt
python-telegram-bot[job-queue]==21.6
python-dotenv==1.0.1
